+++ test/network.cpp
@@ -41,7 +41,7 @@ void test_network()
 
 	std::random_device rd;
 	std::mt19937 gen(rd());
-	std::uniform_real_distribution<double> distr(-0.5, 0.5);
+	std::uniform_real_distribution<float> distr(-0.5, 0.5);
 
 	auto random_values = [&distr, &gen]() { return distr(gen); };
 
@@ -53,12 +53,12 @@ void test_network()
 	auto net = neural_network::make_network(
 
 		neural_network::make_fully_connected_layer<m5x2, m5>(
-			random_values, 0.00003),
+			random_values, 0.00003f),
 
 		neural_network::make_relu_activation_layer<m5>(),
 
 		neural_network::make_fully_connected_layer<m5, m2x2>(
-			random_values, 0.00005),
+			random_values, 0.00005f),
 
 		neural_network::make_logistic_activation_layer<m2x2>()
 	);
@@ -66,11 +66,11 @@ void test_network()
 	m5x2::tensor_type input(random_values);
 	m2x2::tensor_type truth;
 
-	truth(0, 0) = 1.0;
+	truth(0, 0) = 1.0f;
 
 	neural_network::squared_error_loss<m2x2> loss;
 
-	double initialLoss = 0.0, finalLoss = 0.0;
+	float initialLoss = 0.0f, finalLoss = 0.0f;
 
 	train_test_network(net, input, truth, loss, initialLoss, finalLoss);
